serverutils introduces to eliminate duplicities in routes and base endpoints
