+++ server/src/main/scala/za/co/absa/atum/server/api/common/http/ServerUtils.scala
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2021 ABSA Group Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package za.co.absa.atum.server.api.common.http
+
+import sttp.tapir.PublicEndpoint
+import sttp.tapir.ztapir._
+import zio.ZIO
+
+trait ServerUtils {
+
+  protected def createServerEndpoint[I, E, O](
+    endpoint: PublicEndpoint[I, E, O, Any],
+    logic: I => ZIO[HttpEnv.Env, E, O]
+  ): ZServerEndpoint[HttpEnv.Env, Any] = {
+    endpoint.zServerLogic(logic).widen[HttpEnv.Env]
+  }
+
+}
